refactor(config): Pull in more schemas
