+++ src/util/context/schema.rs
@@ -14,6 +14,7 @@ use crate::util::data_structures::HashMap;
 use std::borrow::Cow;
 use std::ffi::OsStr;
 
+use cargo_credential::Secret;
 use serde::Deserialize;
 use serde_untagged::UntaggedEnumVisitor;
 
@@ -21,6 +22,8 @@ use std::path::Path;
 
 use crate::CargoResult;
 
+use super::OptValue;
+use super::PathAndArgs;
 use super::StringList;
 use super::Value;
 use super::path::ConfigRelativePath;
@@ -518,3 +521,56 @@ impl EnvConfigValue {
 }
 
 pub type EnvConfig = HashMap<String, EnvConfigValue>;
+
+/// `[registries.NAME]` tables.
+///
+/// The values here should be kept in sync with `GlobalRegistryConfig`
+#[derive(Deserialize, Clone, Debug)]
+#[serde(rename_all = "kebab-case")]
+pub struct RegistryConfig {
+    pub index: Option<String>,
+    pub token: OptValue<Secret<String>>,
+    pub credential_provider: Option<PathAndArgs>,
+    pub secret_key: OptValue<Secret<String>>,
+    pub secret_key_subject: Option<String>,
+    /// Minimum publish age threshold for RFC 3923
+    pub min_publish_age: Option<String>,
+    #[serde(rename = "protocol")]
+    _protocol: Option<String>,
+}
+
+/// The `[registry]` table, which has more keys than the `[registries.NAME]` tables.
+///
+/// Note: nesting `RegistryConfig` inside this struct and using `serde(flatten)` *should* work
+/// but fails with "invalid type: sequence, expected a value" when attempting to deserialize.
+#[derive(Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub struct GlobalRegistryConfig {
+    pub index: Option<String>,
+    pub token: OptValue<Secret<String>>,
+    pub credential_provider: Option<PathAndArgs>,
+    pub secret_key: OptValue<Secret<String>>,
+    pub secret_key_subject: Option<String>,
+    /// Minimum publish age threshold for RFC 3923
+    pub min_publish_age: Option<String>,
+    /// Global default Minimum publish age threshold for RFC 3923
+    pub global_min_publish_age: Option<String>,
+    #[serde(rename = "default")]
+    _default: Option<String>,
+    #[serde(rename = "global-credential-providers")]
+    _global_credential_providers: Option<Vec<String>>,
+}
+
+impl GlobalRegistryConfig {
+    pub fn to_registry_config(self) -> RegistryConfig {
+        RegistryConfig {
+            index: self.index,
+            token: self.token,
+            credential_provider: self.credential_provider,
+            secret_key: self.secret_key,
+            secret_key_subject: self.secret_key_subject,
+            min_publish_age: self.min_publish_age,
+            _protocol: None,
+        }
+    }
+}
